refactor(chat): split ChatButton.tsx (1676 lines) into focused files
- types.ts: INVITE_PREFIX, ParsedInvite, parseInviteContent
- hooks/useChatInvite.ts: invite state + handlers (send/accept/decline)
- components/ChannelsSidebar.tsx: left sidebar (channels + groups)
- components/DmList.tsx: DM list when no room selected
- components/ConversationHeader.tsx: room header with action buttons
- components/MessageList.tsx: messages + typing indicator
- components/MessageInput.tsx: send form
- components/SendInviteModal.tsx: game invite send modal
- components/AcceptInviteModal.tsx: game invite accept modal
ChatButton.tsx: 1676 → 961 lines (orchestrator only)
