+++ srcs/frontend/src/features/chat/ChatButton.tsx
@@ -7,45 +7,16 @@ import { useTranslation } from "react-i18next";
 import { useChatWebSocket } from "../../hooks/useWebSocket";
 import { getSocket } from "../../hooks/socketSingleton";
 import { setRoomLastSeen, useChatNotifications } from "../../hooks/useChatNotifications";
-import { acceptInvite, cancelInvite, createDmInvite } from "../game/api/matchmaking";
-import { savePendingInvite } from "../../utils/pendingInvite";
-import { markInviteResolved, getResolvedIds, onResolvedUpdated } from "../../utils/resolvedInvites";
 import { getFriendsWithStatus, getBlockedIds } from "../profile/api/friends";
-import { CHARACTERS } from "../../utils/characters";
-import { SCORE_OPTIONS, TIMER_OPTIONS, SCORE_DEFAULT, TIMER_DEFAULT } from "../game/engine/constants";
-import { THEMES, THEME_DEFAULT } from "../game/themes";
-
-// Marker used to identify chat messages that are actually game invites.
-// Format: __GAME_INVITE__|<match_id>|<from_user_id>|<from_name>|<from_nation>|<winning_score>|<duration>|<theme_id>
-const INVITE_PREFIX = "__GAME_INVITE__|";
-
-interface ParsedInvite {
-    matchId: number;
-    fromUserId: number;
-    fromName: string;
-    fromNation: string;
-    winningScore: 3 | 5 | null;
-    duration: 30 | 60 | null;
-    themeId: string;
-}
-
-function parseInviteContent(content: string): ParsedInvite | null {
-    if (!content.startsWith(INVITE_PREFIX)) return null;
-    const parts = content.slice(INVITE_PREFIX.length).split("|");
-    if (parts.length < 4) return null;
-    const matchId = Number(parts[0]);
-    const fromUserId = Number(parts[1]);
-    if (!Number.isFinite(matchId) || matchId <= 0) return null;
-    if (!Number.isFinite(fromUserId) || fromUserId <= 0) return null;
-    const rawScore = parts[4];
-    const rawDuration = parts[5];
-    const winningScore = rawScore === "null" ? null : ([3, 5] as const).includes(Number(rawScore) as 3 | 5) ? Number(rawScore) as 3 | 5 : SCORE_DEFAULT;
-    const duration = rawDuration === "null" ? null : ([30, 60] as const).includes(Number(rawDuration) as 30 | 60) ? Number(rawDuration) as 30 | 60 : TIMER_DEFAULT;
-    const themeId = parts[6] || THEME_DEFAULT.id;
-    return { matchId, fromUserId, fromName: parts[2], fromNation: parts[3], winningScore, duration, themeId };
-}
-
-const TYPING_TIMEOUT_MS = 3500;
+import { TYPING_TIMEOUT_MS } from "./types";
+import { useChatInvite } from "./hooks/useChatInvite";
+import { ChannelsSidebar } from "./components/ChannelsSidebar";
+import { DmList } from "./components/DmList";
+import { ConversationHeader } from "./components/ConversationHeader";
+import { MessageList } from "./components/MessageList";
+import { MessageInput } from "./components/MessageInput";
+import { SendInviteModal } from "./components/SendInviteModal";
+import { AcceptInviteModal } from "./components/AcceptInviteModal";
 
 type ChatButtonProps = {
     initialRoomId?: number | null;
@@ -79,25 +50,13 @@ export function ChatButton({ initialRoomId = null, initialDmUserId = null }: Cha
     const [isChannelsOpen, setIsChannelsOpen] = useState(false);
     const [typingUser, setTypingUser] = useState<{ userId: number; username?: string } | null>(null);
     const [partnerLastReadAt, setPartnerLastReadAt] = useState<string | null>(null);
-    const [isInviting, setIsInviting] = useState(false);
-    const [busyInviteId, setBusyInviteId] = useState<number | null>(null);
-    const [showInviteModal, setShowInviteModal] = useState(false);
-    const [inviteNation, setInviteNation] = useState<string>("Algeria");
-    const [inviteScore, setInviteScore] = useState<3 | 5 | null>(SCORE_DEFAULT);
-    const [inviteDuration, setInviteDuration] = useState<30 | 60 | null>(TIMER_DEFAULT);
-    const [inviteThemeId, setInviteThemeId] = useState<string>(THEME_DEFAULT.id);
-    const [pendingAccept, setPendingAccept] = useState<ParsedInvite | null>(null);
-    const [acceptNation, setAcceptNation] = useState<string>("Algeria");
-    const [resolvedInviteIds, setResolvedInviteIds] = useState<Set<number>>(() => getResolvedIds());
+
     const avatarBlobsRef = useRef<Record<number, string>>({});
     const seenMemberEventRef = useRef<Set<string>>(new Set());
     const membersLoadedRoomIdRef = useRef<number | null>(null);
     const messageEndRef = useRef<HTMLDivElement | null>(null);
     const lastTypingSentRef = useRef<number>(0);
     const typingClearTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-
-    useEffect(() => onResolvedUpdated(setResolvedInviteIds), [])
 
     const { t } = useTranslation();
 
@@ -134,17 +93,6 @@ export function ChatButton({ initialRoomId = null, initialDmUserId = null }: Cha
         if (!otherUserId) return room.name;
         const partner = profiles[otherUserId];
         return partner?.display_name?.trim() || "Direct message";
-    };
-
-    const getUserDisplayName = (userId: number) => {
-        if (userId <= 0) return t("Anonymous");
-        const profile = profiles[userId];
-        return profile?.display_name?.trim() || `User ${userId}`;
-    };
-
-    const getUserAvatarUrl = (userId: number) => {
-        if (userId <= 0) return "/assets/default-profile.jpg";
-        return renderAvatar(userId);
     };
 
     const visibleRooms = useMemo(() => {
@@ -188,6 +136,15 @@ export function ChatButton({ initialRoomId = null, initialDmUserId = null }: Cha
     const [onlineIds, setOnlineIds] = useState<Set<number>>(new Set());
     const [allFriendIds, setAllFriendIds] = useState<number[]>([]);
     const [isNewDmOpen, setIsNewDmOpen] = useState(false);
+
+    const invite = useChatInvite({
+        token,
+        currentUserId,
+        isSelectedRoomDm,
+        selectedDmUserId,
+        profiles,
+        sendWsMessage,
+    });
 
     useEffect(() => {
         if (!token) return;
@@ -649,88 +606,6 @@ export function ChatButton({ initialRoomId = null, initialDmUserId = null }: Cha
         }
     };
 
-    const handleSendInvite = () => {
-        if (!isSelectedRoomDm || !selectedDmUserId || !token) return;
-        const me = profiles[currentUserId];
-        setInviteNation(me?.country?.trim() || "Algeria");
-        setInviteScore(SCORE_DEFAULT);
-        setInviteDuration(TIMER_DEFAULT);
-        setInviteThemeId(THEME_DEFAULT.id);
-        setShowInviteModal(true);
-    };
-
-    const handleConfirmInvite = async () => {
-        if (!isSelectedRoomDm || !selectedDmUserId || !token) return;
-        const myName = localStorage.getItem("username") || "Player";
-        try {
-            setIsInviting(true);
-            setShowInviteModal(false);
-            const result = await createDmInvite({
-                targetUserId: selectedDmUserId,
-                playerName: myName,
-                playerNation: inviteNation,
-                winningScore: inviteScore,
-                duration: inviteDuration,
-            });
-            if (result.status !== "matched" || !result.match_id) {
-                setStatus(t("Failed to send invite"));
-                return;
-            }
-            const inviteContent = `${INVITE_PREFIX}${result.match_id}|${currentUserId}|${myName}|${inviteNation}|${inviteScore}|${inviteDuration}|${inviteThemeId}`;
-            sendWsMessage(inviteContent);
-            savePendingInvite({ ...result, myNation: inviteNation, themeId: inviteThemeId });
-            setStatus(t("Invite sent! Waiting for opponent…"));
-        } catch (err) {
-            setStatus(err instanceof Error ? err.message : t("Failed to send invite"));
-        } finally {
-            setIsInviting(false);
-        }
-    };
-
-    const handleAcceptInvite = (invite: ParsedInvite) => {
-        const me = profiles[currentUserId];
-        setAcceptNation(me?.country?.trim() || "Algeria");
-        setPendingAccept(invite);
-    };
-
-    const handleConfirmAccept = async () => {
-        if (!pendingAccept || !token) return;
-        const invite = pendingAccept;
-        const myName = localStorage.getItem("username") || "Player";
-        try {
-            setBusyInviteId(invite.matchId);
-            setPendingAccept(null);
-            const result = await acceptInvite({
-                matchId: invite.matchId,
-                playerName: myName,
-                playerNation: acceptNation,
-            });
-            if (result.status !== "matched" || !result.match_id) {
-                markInviteResolved(invite.matchId);
-                setStatus(t("Invite expired"));
-                return;
-            }
-            markInviteResolved(invite.matchId);
-            navigate("/online-gameplay", { state: { ...result, myNation: acceptNation, themeId: invite.themeId } });
-        } catch {
-            markInviteResolved(invite.matchId);
-        } finally {
-            setBusyInviteId(null);
-        }
-    };
-
-    const handleDeclineInvite = async (invite: ParsedInvite) => {
-        try {
-            setBusyInviteId(invite.matchId);
-            await cancelInvite(invite.matchId);
-        } catch {
-            // best-effort
-        } finally {
-            markInviteResolved(invite.matchId);
-            setBusyInviteId(null);
-        }
-    };
-
     const handleSendMessage = async () => {
         if (!token || selectedRoomId == null) return;
         if (!messageText.trim()) return;
@@ -927,414 +802,84 @@ export function ChatButton({ initialRoomId = null, initialDmUserId = null }: Cha
                         className="fixed inset-0 z-30 bg-black/40 lg:hidden"
                     />
                 )}
-                <aside className={`fixed inset-y-0 left-0 z-40 flex w-[min(86vw,19rem)] flex-col border-r-4 border-[#1f2937] bg-[#ece3d0] shadow-[10px_0_0_#1f2937] transition-transform duration-200 lg:static lg:z-auto lg:w-auto lg:translate-x-0 lg:shadow-none ${isChannelsOpen ? "translate-x-0" : "-translate-x-full lg:translate-x-0"}`}>
-                    {/* Sidebar header */}
-                    <div className="border-b-2 border-[#1f2937]/15 px-3 py-3 flex items-center justify-between shrink-0">
-                        <span className="text-xs font-bold uppercase tracking-[0.18em] text-[#374151]">{t("Rooms")}</span>
-                        <button
-                            type="button"
-                            onClick={() => setIsChannelsOpen(false)}
-                            className="rounded-full border-2 border-[#1f2937] bg-white px-2 py-1 text-[11px] font-semibold uppercase tracking-[0.16em] text-[#1f2937] lg:hidden"
-                        >{t("Close")}</button>
-                    </div>
-
-                    <div className="min-h-0 flex-1 overflow-auto px-3 py-3 flex flex-col gap-4">
-
-                        {/* ── Public channels ── */}
-                        <section>
-                            <div className="flex items-center justify-between mb-2">
-                                <span className="text-[10px] font-bold uppercase tracking-[0.2em] text-[#9ca3af]"># {t("Public channels")}</span>
-                                <button
-                                    type="button"
-                                    onClick={() => { setIsCreateFormOpen(v => !v); setIsCreateGroupOpen(false); }}
-                                    title={t("Create channel")}
-                                    className="rounded-full border-2 border-[#1f2937] bg-[#4AD95A] w-6 h-6 flex items-center justify-center text-xs font-bold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px]"
-                                >+</button>
-                            </div>
-                            {isCreateFormOpen && (
-                                <div className="flex gap-2 mb-2">
-                                    <input
-                                        type="text"
-                                        placeholder={t("Channel name")}
-                                        value={roomName}
-                                        onChange={(e) => setRoomName(e.target.value)}
-                                        onKeyDown={(e) => { if (e.key === 'Enter') { e.preventDefault(); handleCreateRoom(); } }}
-                                        className="min-w-0 flex-1 rounded-xl border-2 border-[#1f2937] bg-white px-3 py-1.5 text-sm outline-none transition focus:border-blue-500"
-                                    />
-                                    <button
-                                        onClick={handleCreateRoom}
-                                        disabled={isCreatingRoom}
-                                        className="rounded-xl border-2 border-[#1f2937] bg-[#4AD95A] px-3 py-1.5 text-sm font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px] disabled:opacity-60"
-                                    >{isCreatingRoom ? "..." : t("Create")}</button>
-                                </div>
-                            )}
-                            <div className="space-y-1.5">
-                                {publicChannels.length === 0 && (
-                                    <p className="px-2 py-2 text-xs text-[#9ca3af]">{t("No public channels yet.")}</p>
-                                )}
-                                {publicChannels.map((room) => {
-                                    const isSelected = room.id === selectedRoomId;
-                                    const hasUnreadRoom = unreadRoomIds.includes(room.id);
-                                    const lastTs = lastMessageByRoomId[room.id];
-                                    const relativeTime = lastTs ? (() => {
-                                        const diffMin = Math.floor((Date.now() - lastTs) / 60000);
-                                        if (diffMin < 1) return t("just now");
-                                        if (diffMin < 60) return `${diffMin}${t("m")}`;
-                                        const diffH = Math.floor(diffMin / 60);
-                                        if (diffH < 24) return `${diffH}h`;
-                                        return `${Math.floor(diffH / 24)}j`;
-                                    })() : null;
-                                    return (
-                                        <button
-                                            key={room.id}
-                                            onClick={() => { setSelectedRoomId(room.id); setIsChannelsOpen(false); }}
-                                            className={`w-full rounded-2xl border-2 px-3 py-2.5 text-left transition ${isSelected ? "border-[#1f2937] bg-white shadow-[3px_3px_0_#1f2937]" : "border-[#1f2937]/20 bg-white/60 hover:bg-white/85"}`}
-                                        >
-                                            <div className="flex items-center gap-2">
-                                                <div className="shrink-0 h-8 w-8 rounded-full border-2 border-[#1f2937] bg-[#1f2937] flex items-center justify-center text-white text-xs font-bold">#</div>
-                                                <div className="min-w-0 flex-1">
-                                                    <div className="flex items-center justify-between gap-1">
-                                                        <div className="truncate text-sm font-semibold text-[#1f2937]">{room.name}</div>
-                                                        <div className="flex items-center gap-1 shrink-0">
-                                                            {hasUnreadRoom && !isSelected && <span className="inline-flex h-2 w-2 rounded-full bg-red-500" />}
-                                                            {relativeTime && <span className="text-[10px] text-[#9ca3af]">{relativeTime}</span>}
-                                                        </div>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                        </button>
-                                    );
-                                })}
-                            </div>
-                        </section>
-
-                        {/* ── Private groups ── */}
-                        <section>
-                            <div className="flex items-center justify-between mb-2">
-                                <span className="text-[10px] font-bold uppercase tracking-[0.2em] text-[#9ca3af]">👥 {t("Groups")}</span>
-                                <button
-                                    type="button"
-                                    onClick={() => { setIsCreateGroupOpen(v => !v); setIsCreateFormOpen(false); }}
-                                    title={t("Create group")}
-                                    className="rounded-full border-2 border-[#1f2937] bg-[#818cf8] w-6 h-6 flex items-center justify-center text-xs font-bold text-white shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px]"
-                                >+</button>
-                            </div>
-                            {isCreateGroupOpen && (
-                                <div className="mb-2 rounded-2xl border-2 border-[#1f2937]/20 bg-white/60 p-3 flex flex-col gap-2">
-                                    <input
-                                        type="text"
-                                        placeholder={t("Group name")}
-                                        value={groupName}
-                                        onChange={(e) => setGroupName(e.target.value)}
-                                        className="rounded-xl border-2 border-[#1f2937] bg-white px-3 py-1.5 text-sm outline-none transition focus:border-blue-500 w-full"
-                                    />
-                                    <div className="text-[10px] font-bold uppercase tracking-[0.16em] text-[#374151]">{t("Invite friends")}</div>
-                                    <div className="flex flex-wrap gap-1.5">
-                                        {allFriendIds.length === 0 && (
-                                            <p className="text-xs text-[#9ca3af]">{t("No friends yet.")}</p>
-                                        )}
-                                        {allFriendIds.map(fid => {
-                                            const name = profiles[fid]?.display_name || `User ${fid}`;
-                                            const sel = groupSelectedFriends.has(fid);
-                                            return (
-                                                <button
-                                                    key={fid}
-                                                    type="button"
-                                                    onClick={() => setGroupSelectedFriends(prev => {
-                                                        const next = new Set(prev);
-                                                        if (sel) next.delete(fid); else next.add(fid);
-                                                        return next;
-                                                    })}
-                                                    className={`flex items-center gap-1.5 rounded-full border-2 px-2.5 py-1 text-xs font-semibold transition ${sel ? "border-[#818cf8] bg-[#818cf8] text-white" : "border-[#1f2937]/20 bg-white text-[#1f2937] hover:border-[#818cf8]"}`}
-                                                >
-                                                    <img src={renderAvatar(fid)} alt={name} onError={handleAvatarError} className="h-4 w-4 rounded-full border border-[#1f2937]/20 object-cover" />
-                                                    {name}
-                                                    {sel && <span className="ml-0.5">✓</span>}
-                                                </button>
-                                            );
-                                        })}
-                                    </div>
-                                    <button
-                                        onClick={handleCreateGroup}
-                                        disabled={isCreatingGroup || !groupName.trim()}
-                                        className="rounded-xl border-2 border-[#1f2937] bg-[#818cf8] px-3 py-1.5 text-sm font-semibold text-white shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px] disabled:opacity-60 w-full"
-                                    >{isCreatingGroup ? "..." : t("Create group")}</button>
-                                </div>
-                            )}
-                            <div className="space-y-1.5">
-                                {privateGroups.length === 0 && (
-                                    <p className="px-2 py-2 text-xs text-[#9ca3af]">{t("No groups yet.")}</p>
-                                )}
-                                {privateGroups.map((room) => {
-                                    const isSelected = room.id === selectedRoomId;
-                                    const hasUnreadRoom = unreadRoomIds.includes(room.id);
-                                    const lastTs = lastMessageByRoomId[room.id];
-                                    const relativeTime = lastTs ? (() => {
-                                        const diffMin = Math.floor((Date.now() - lastTs) / 60000);
-                                        if (diffMin < 1) return t("just now");
-                                        if (diffMin < 60) return `${diffMin}${t("m")}`;
-                                        const diffH = Math.floor(diffMin / 60);
-                                        if (diffH < 24) return `${diffH}h`;
-                                        return `${Math.floor(diffH / 24)}j`;
-                                    })() : null;
-                                    return (
-                                        <button
-                                            key={room.id}
-                                            onClick={() => { setSelectedRoomId(room.id); setIsChannelsOpen(false); }}
-                                            className={`w-full rounded-2xl border-2 px-3 py-2.5 text-left transition ${isSelected ? "border-[#1f2937] bg-white shadow-[3px_3px_0_#1f2937]" : "border-[#1f2937]/20 bg-white/60 hover:bg-white/85"}`}
-                                        >
-                                            <div className="flex items-center gap-2">
-                                                <div className="shrink-0 h-8 w-8 rounded-full border-2 border-[#818cf8] bg-[#818cf8] flex items-center justify-center text-white text-xs font-bold">👥</div>
-                                                <div className="min-w-0 flex-1">
-                                                    <div className="flex items-center justify-between gap-1">
-                                                        <div className="truncate text-sm font-semibold text-[#1f2937]">{room.name}</div>
-                                                        <div className="flex items-center gap-1 shrink-0">
-                                                            {hasUnreadRoom && !isSelected && <span className="inline-flex h-2 w-2 rounded-full bg-red-500" />}
-                                                            {relativeTime && <span className="text-[10px] text-[#9ca3af]">{relativeTime}</span>}
-                                                        </div>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                        </button>
-                                    );
-                                })}
-                            </div>
-                        </section>
-                    </div>
-                </aside>
+                <ChannelsSidebar
+                    isChannelsOpen={isChannelsOpen}
+                    setIsChannelsOpen={setIsChannelsOpen}
+                    isCreateFormOpen={isCreateFormOpen}
+                    setIsCreateFormOpen={setIsCreateFormOpen}
+                    isCreateGroupOpen={isCreateGroupOpen}
+                    setIsCreateGroupOpen={setIsCreateGroupOpen}
+                    roomName={roomName}
+                    setRoomName={setRoomName}
+                    isCreatingRoom={isCreatingRoom}
+                    handleCreateRoom={handleCreateRoom}
+                    publicChannels={publicChannels}
+                    privateGroups={privateGroups}
+                    selectedRoomId={selectedRoomId}
+                    setSelectedRoomId={setSelectedRoomId}
+                    unreadRoomIds={unreadRoomIds}
+                    lastMessageByRoomId={lastMessageByRoomId}
+                    allFriendIds={allFriendIds}
+                    profiles={profiles}
+                    groupName={groupName}
+                    setGroupName={setGroupName}
+                    groupSelectedFriends={groupSelectedFriends}
+                    setGroupSelectedFriends={setGroupSelectedFriends}
+                    isCreatingGroup={isCreatingGroup}
+                    handleCreateGroup={handleCreateGroup}
+                    renderAvatar={renderAvatar}
+                    handleAvatarError={handleAvatarError}
+                />
 
                 {/* ── Main area: DM list or conversation ── */}
                 <main className="flex min-h-0 flex-col bg-[#f7f3ea] lg:col-span-1">
 
                     {/* DM list — shown when no room is selected (all screens) */}
                     {!selectedRoom && (
-                        <div className="flex min-h-0 flex-col flex-1">
-                            <div className="border-b-2 border-[#1f2937]/15 px-4 py-3 flex items-center justify-between">
-                                <span className="text-xs font-bold uppercase tracking-[0.18em] text-[#374151]">{t("Direct Messages")}</span>
-                                <button
-                                    type="button"
-                                    onClick={() => { setIsNewDmOpen(v => !v); }}
-                                    title={t("New conversation")}
-                                    className="rounded-full border-2 border-[#1f2937] bg-[#4AD95A] w-7 h-7 flex items-center justify-center text-sm font-bold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px]"
-                                >+</button>
-                            </div>
-                            {/* New DM friend picker */}
-                            {isNewDmOpen && (
-                                <div className="border-b-2 border-[#1f2937]/15 bg-[#ece3d0] px-3 py-3">
-                                    <div className="text-[11px] font-bold uppercase tracking-[0.16em] text-[#374151] mb-2">{t("Start a conversation")}</div>
-                                    {friendsWithoutDm.length === 0 ? (
-                                        <p className="text-xs text-[#6b7280] px-1">{t("All friends already have a conversation.")}</p>
-                                    ) : (
-                                        <div className="flex flex-wrap gap-2">
-                                            {friendsWithoutDm.map(friendId => {
-                                                const profile = profiles[friendId];
-                                                const name = profile?.display_name || `User ${friendId}`;
-                                                const isOnline = onlineIds.has(friendId);
-                                                return (
-                                                    <button
-                                                        key={friendId}
-                                                        type="button"
-                                                        onClick={() => handleStartDm(friendId)}
-                                                        className="flex items-center gap-2 rounded-2xl border-2 border-[#1f2937]/20 bg-white/80 px-3 py-2 text-left text-sm font-semibold text-[#1f2937] transition hover:bg-white hover:border-[#1f2937]"
-                                                    >
-                                                        <div className="relative shrink-0">
-                                                            <img
-                                                                src={renderAvatar(friendId)}
-                                                                alt={name}
-                                                                onError={handleAvatarError}
-                                                                className="h-8 w-8 rounded-full border-2 border-[#1f2937] object-cover"
-                                                            />
-                                                            {isOnline && <span className="absolute bottom-0 right-0 h-2 w-2 rounded-full bg-green-500 border border-white" />}
-                                                        </div>
-                                                        <span className="truncate max-w-[8rem]">{name}</span>
-                                                    </button>
-                                                );
-                                            })}
-                                        </div>
-                                    )}
-                                </div>
-                            )}
-                            <div className="min-h-0 flex-1 overflow-auto px-3 py-3">
-                                <div className="space-y-2">
-                                    {dmRooms.length === 0 && (
-                                        <p className="px-2 py-6 text-center text-sm text-[#6b7280]">{t("No direct messages yet.")}</p>
-                                    )}
-                                    {dmRooms.map((room) => {
-                                        const dmUserId = getDmOtherUserId(room)!;
-                                        const hasUnreadRoom = unreadRoomIds.includes(room.id);
-                                        const isOnline = onlineIds.has(dmUserId);
-                                        const lastTs = lastMessageByRoomId[room.id];
-                                        const relativeTime = lastTs ? (() => {
-                                            const diffMin = Math.floor((Date.now() - lastTs) / 60000);
-                                            if (diffMin < 1) return t("just now", "just now");
-                                            if (diffMin < 60) return `${diffMin} ${t("min ago", "min ago")}`;
-                                            const diffH = Math.floor(diffMin / 60);
-                                            if (diffH < 24) return `${diffH}h ${t("ago", "ago")}`;
-                                            return `${Math.floor(diffH / 24)}d ${t("ago", "ago")}`;
-                                        })() : null;
-                                        return (
-                                            <button
-                                                key={room.id}
-                                                onClick={() => setSelectedRoomId(room.id)}
-                                                className="w-full rounded-2xl border-2 border-[#1f2937]/20 bg-white/70 px-3 py-3 text-left transition hover:bg-white"
-                                            >
-                                                <div className="flex items-center gap-3">
-                                                    <div className="relative shrink-0">
-                                                        <img
-                                                            src={renderAvatar(dmUserId)}
-                                                            alt={getRoomDisplayName(room)}
-                                                            onError={handleAvatarError}
-                                                            className="h-11 w-11 rounded-full border-2 border-[#1f2937] object-cover"
-                                                        />
-                                                        {isOnline && <span className="absolute bottom-0 right-0 h-3 w-3 rounded-full bg-green-500 border-2 border-white" />}
-                                                    </div>
-                                                    <div className="min-w-0 flex-1">
-                                                        <div className="flex items-center justify-between gap-1">
-                                                            <span className="truncate text-sm font-semibold text-[#1f2937]">{getRoomDisplayName(room)}</span>
-                                                            <div className="flex items-center gap-1.5 shrink-0">
-                                                                {hasUnreadRoom && <span className="inline-flex h-2.5 w-2.5 rounded-full bg-red-500" />}
-                                                                {relativeTime && <span className="text-[10px] text-[#9ca3af]">{relativeTime}</span>}
-                                                            </div>
-                                                        </div>
-                                                        <span className={`text-xs font-medium ${isOnline ? "text-green-600" : "text-[#9ca3af]"}`}>
-                                                            {isOnline ? t("Online") : t("Offline")}
-                                                        </span>
-                                                    </div>
-                                                </div>
-                                            </button>
-                                        );
-                                    })}
-                                </div>
-                            </div>
-                        </div>
+                        <DmList
+                            isNewDmOpen={isNewDmOpen}
+                            setIsNewDmOpen={setIsNewDmOpen}
+                            friendsWithoutDm={friendsWithoutDm}
+                            profiles={profiles}
+                            onlineIds={onlineIds}
+                            handleStartDm={handleStartDm}
+                            dmRooms={dmRooms}
+                            getDmOtherUserId={getDmOtherUserId}
+                            getRoomDisplayName={getRoomDisplayName}
+                            unreadRoomIds={unreadRoomIds}
+                            lastMessageByRoomId={lastMessageByRoomId}
+                            setSelectedRoomId={setSelectedRoomId}
+                            renderAvatar={renderAvatar}
+                            handleAvatarError={handleAvatarError}
+                        />
                     )}
 
-                    {/* Conversation header — hidden when no room selected */}
-                    <div className={`border-b-4 border-[#1f2937] bg-white/70 px-3 py-3 backdrop-blur-sm min-[481px]:px-4 ${!selectedRoom ? "hidden" : ""}`}>
-                        {selectedRoom ? (
-                            <div className="flex items-center justify-between gap-3">
-                                <div className="flex items-center gap-3 min-w-0">
-                                    {/* Back button on mobile */}
-                                    <button
-                                        type="button"
-                                        onClick={() => setSelectedRoomId(null)}
-                                        className="shrink-0 lg:hidden rounded-full border-2 border-[#1f2937] bg-white w-8 h-8 flex items-center justify-center text-[#1f2937] font-bold shadow-[2px_2px_0_#1f2937]"
-                                        aria-label="Back"
-                                    >←</button>
-                                    {isSelectedRoomDm && selectedDmUserId ? (
-                                        <div className="relative shrink-0">
-                                            <img
-                                                src={renderAvatar(selectedDmUserId)}
-                                                alt={getRoomDisplayName(selectedRoom)}
-                                                onError={handleAvatarError}
-                                                className="h-10 w-10 rounded-full border-2 border-[#1f2937] object-cover"
-                                            />
-                                            {onlineIds.has(selectedDmUserId) && (
-                                                <span className="absolute bottom-0 right-0 h-3 w-3 rounded-full bg-green-500 border-2 border-white" />
-                                            )}
-                                        </div>
-                                    ) : selectedRoom?.is_private ? (
-                                        <div className="shrink-0 h-10 w-10 rounded-full border-2 border-[#818cf8] bg-[#818cf8] flex items-center justify-center text-white font-bold text-base">
-                                            👥
-                                        </div>
-                                    ) : (
-                                        <div className="shrink-0 h-10 w-10 rounded-full border-2 border-[#1f2937] bg-[#1f2937] flex items-center justify-center text-white font-bold text-base">
-                                            #
-                                        </div>
-                                    )}
-                                    <div className="min-w-0">
-                                        <div className="truncate text-base font-bold text-[#1f2937]">{getRoomDisplayName(selectedRoom)}</div>
-                                        {isSelectedRoomDm && selectedDmUserId ? (
-                                            <div className={`text-xs font-medium ${onlineIds.has(selectedDmUserId) ? "text-green-600" : "text-[#9ca3af]"}`}>
-                                                {onlineIds.has(selectedDmUserId) ? t("Online") : t("Offline")}
-                                            </div>
-                                        ) : (
-                                            <div className="text-xs text-[#6b7280]">
-                                                {memberIds.length} {t("member")}{memberIds.length > 1 ? t("s") : ""}
-                                            </div>
-                                        )}
-                                    </div>
-                                </div>
-
-                                <div className="flex items-center gap-1.5 shrink-0">
-                                    {/* Invite member button — private groups only (any member can invite) */}
-                                    {selectedRoom?.is_private && !isSelectedRoomDm && isMember && (
-                                        <div className="relative">
-                                            <button
-                                                type="button"
-                                                onClick={() => setIsInviteMemberOpen(v => !v)}
-                                                title={t("Invite a friend")}
-                                                className="rounded-xl border-2 border-[#1f2937] bg-[#818cf8] px-2.5 py-2 text-sm font-bold text-white shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px]"
-                                            >+ 👤</button>
-                                            {isInviteMemberOpen && (
-                                                <div className="absolute right-0 top-full mt-2 z-50 w-56 rounded-2xl border-2 border-[#1f2937] bg-white shadow-[4px_4px_0_#1f2937] p-3 flex flex-col gap-2">
-                                                    <div className="text-[10px] font-bold uppercase tracking-[0.16em] text-[#374151]">{t("Invite a friend")}</div>
-                                                    {allFriendIds.filter(id => !roomMemberIds.includes(id)).length === 0 ? (
-                                                        <p className="text-xs text-[#9ca3af]">{t("All friends are already in this group.")}</p>
-                                                    ) : (
-                                                        allFriendIds.filter(id => !roomMemberIds.includes(id)).map(fid => {
-                                                            const name = profiles[fid]?.display_name || `User ${fid}`;
-                                                            return (
-                                                                <button
-                                                                    key={fid}
-                                                                    type="button"
-                                                                    disabled={isInvitingMember}
-                                                                    onClick={() => handleInviteMember(fid)}
-                                                                    className="flex items-center gap-2 rounded-xl border-2 border-[#1f2937]/20 bg-[#f5efe2] px-3 py-1.5 text-sm font-semibold text-[#1f2937] transition hover:bg-[#ece3d0] text-left"
-                                                                >
-                                                                    <img src={renderAvatar(fid)} alt={name} onError={handleAvatarError} className="h-6 w-6 rounded-full border border-[#1f2937]/20 object-cover shrink-0" />
-                                                                    <span className="truncate">{name}</span>
-                                                                </button>
-                                                            );
-                                                        })
-                                                    )}
-                                                </div>
-                                            )}
-                                        </div>
-                                    )}
-                                    {isSelectedRoomDm && (
-                                        <button
-                                            type="button"
-                                            onClick={handleSendInvite}
-                                            disabled={isInviting}
-                                            title={isInviting ? t("Sending...") : t("Invite to game")}
-                                            className="rounded-xl border-2 border-[#1f2937] bg-[#facc15] px-2.5 py-2 text-base font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px] disabled:cursor-not-allowed disabled:opacity-60"
-                                        >
-                                            🎮
-                                        </button>
-                                    )}
-                                    {!isSelectedRoomDm && !isMember && (
-                                        <button
-                                            onClick={handleJoinRoom}
-                                            className="rounded-xl border-2 border-[#1f2937] bg-[#4AD95A] px-2.5 py-1.5 text-xs font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px]"
-                                        >
-                                            {t("Join")}
-                                        </button>
-                                    )}
-                                    {!isSelectedRoomDm && isMember && (
-                                        <button
-                                            onClick={handleLeaveRoom}
-                                            className="rounded-xl border-2 border-[#1f2937] bg-white px-2.5 py-1.5 text-xs font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px]"
-                                        >
-                                            {t("Leave")}
-                                        </button>
-                                    )}
-                                    {isOwner && !isSelectedRoomDm && (
-                                        <button
-                                            onClick={handleDeleteRoom}
-                                            className="rounded-xl border-2 border-[#1f2937] bg-[#ef4444] px-2.5 py-1.5 text-xs font-semibold text-white shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px]"
-                                        >
-                                            {t("Delete")}
-                                        </button>
-                                    )}
-                                </div>
-                            </div>
-                        ) : (
-                            <div className="py-1 hidden lg:block">
-                                <div className="text-lg font-bold text-[#1f2937]">{t("Select a conversation")}</div>
-                            </div>
-                        )}
-                    </div>
+                    {/* Conversation header */}
+                    <ConversationHeader
+                        selectedRoom={selectedRoom}
+                        isSelectedRoomDm={isSelectedRoomDm}
+                        selectedDmUserId={selectedDmUserId}
+                        onlineIds={onlineIds}
+                        memberIds={memberIds}
+                        roomMemberIds={roomMemberIds}
+                        allFriendIds={allFriendIds}
+                        profiles={profiles}
+                        isMember={isMember}
+                        isOwner={isOwner}
+                        isInviting={invite.isInviting}
+                        isInviteMemberOpen={isInviteMemberOpen}
+                        setIsInviteMemberOpen={setIsInviteMemberOpen}
+                        isInvitingMember={isInvitingMember}
+                        getRoomDisplayName={getRoomDisplayName}
+                        renderAvatar={renderAvatar}
+                        handleAvatarError={handleAvatarError}
+                        handleSendInvite={invite.handleSendInvite}
+                        handleJoinRoom={handleJoinRoom}
+                        handleLeaveRoom={handleLeaveRoom}
+                        handleDeleteRoom={handleDeleteRoom}
+                        handleInviteMember={handleInviteMember}
+                        setSelectedRoomId={setSelectedRoomId}
+                    />
 
                     {status && (
                         <div className={`border-b-4 border-[#1f2937] bg-[#fff7d6] px-4 py-2.5 text-sm font-medium text-[#1f2937] ${!selectedRoom ? "hidden" : ""}`}>
@@ -1342,335 +887,75 @@ export function ChatButton({ initialRoomId = null, initialDmUserId = null }: Cha
                         </div>
                     )}
 
-                    {/* Messages area — hidden when no room selected */}
-                    <div className={`min-h-0 flex-1 overflow-auto px-3 py-3 min-[481px]:px-4 min-[481px]:py-4 ${!selectedRoom ? "hidden" : ""}`}>
-                        <div className="flex min-h-full flex-col justify-end gap-3">
-                            {messages.length === 0 && systemEvents.length === 0 && (
-                                <div className="rounded-2xl border-2 border-dashed border-[#1f2937]/20 bg-white/60 px-4 py-6 text-center text-sm text-[#6b7280]">
-                                    {t("No messages yet in this room.")}
-                                </div>
-                            )}
+                    {/* Messages area */}
+                    <MessageList
+                        selectedRoom={Boolean(selectedRoom)}
+                        messages={messages}
+                        groupedMessages={groupedMessages}
+                        systemEvents={systemEvents}
+                        profiles={profiles}
+                        currentUserId={currentUserId}
+                        isSelectedRoomDm={isSelectedRoomDm}
+                        partnerLastReadAt={partnerLastReadAt}
+                        isMember={isMember}
+                        typingUser={typingUser}
+                        resolvedInviteIds={invite.resolvedInviteIds}
+                        busyInviteId={invite.busyInviteId}
+                        messageEndRef={messageEndRef}
+                        formatMsgTime={formatMsgTime}
+                        renderAvatar={renderAvatar}
+                        handleAvatarError={handleAvatarError}
+                        openProfile={openProfile}
+                        handleAcceptInvite={invite.handleAcceptInvite}
+                        handleDeclineInvite={invite.handleDeclineInvite}
+                    />
 
-                            {groupedMessages.map((group) => (
-                                <div key={group.dateKey} className="flex flex-col gap-3">
-                                    {/* Date separator */}
-                                    {group.dateLabel && (
-                                        <div className="flex items-center gap-2 my-1">
-                                            <div className="flex-1 h-px bg-[#1f2937]/10" />
-                                            <span className="rounded-full border border-[#1f2937]/15 bg-white/80 px-3 py-0.5 text-[11px] font-semibold uppercase tracking-[0.16em] text-[#6b7280]">
-                                                {group.dateLabel}
-                                            </span>
-                                            <div className="flex-1 h-px bg-[#1f2937]/10" />
-                                        </div>
-                                    )}
-
-                                    {group.messages.map(({ msg: message, idx: index }, posInGroup) => {
-                                        const profile = profiles[message.sender_user_id];
-                                        const displayName = profile?.display_name || `User ${message.sender_user_id}`;
-                                        const avatarUrl = renderAvatar(message.sender_user_id);
-                                        const isMine = currentUserId > 0 && message.sender_user_id === currentUserId;
-                                        const invite = parseInviteContent(message.content);
-                                        const isLastMine = isMine && index === messages.length - 1;
-                                        const seenByPartner = Boolean(
-                                            isLastMine && isSelectedRoomDm && partnerLastReadAt && message.created_at &&
-                                            Date.parse(partnerLastReadAt) >= Date.parse(message.created_at)
-                                        );
-                                        const nextItem = group.messages[posInGroup + 1];
-                                        const isLastInRun = !nextItem || nextItem.msg.sender_user_id !== message.sender_user_id;
-                                        const showAvatar = isLastInRun;
-
-                                        return (
-                                            <article
-                                                key={`${message.room_id}-${message.sender_user_id}-${message.created_at ?? index}`}
-                                                className={`flex items-end gap-2 sm:gap-3 ${isMine ? "justify-end" : "justify-start"}`}
-                                            >
-                                                {!isMine && (
-                                                    <div className="shrink-0 w-9 sm:w-10">
-                                                        {showAvatar ? (
-                                                            <button
-                                                                type="button"
-                                                                onClick={() => openProfile(message.sender_user_id)}
-                                                                aria-label={`Open ${displayName}'s profile`}
-                                                            >
-                                                                <img
-                                                                    src={avatarUrl}
-                                                                    alt={displayName}
-                                                                    onError={handleAvatarError}
-                                                                    className="h-9 w-9 rounded-full border-2 border-[#1f2937] object-cover sm:h-10 sm:w-10"
-                                                                />
-                                                            </button>
-                                                        ) : null}
-                                                    </div>
-                                                )}
-
-                                                <div className="flex max-w-[86%] flex-col gap-1 sm:max-w-[80%]">
-                                                    {invite ? (
-                                                        <div
-                                                            className={`rounded-3xl border-2 border-[#1f2937] px-3 py-3 shadow-[4px_4px_0_#1f2937] sm:px-4 sm:py-4 ${
-                                                                isMine ? "bg-blue-700 text-white" : "bg-white text-[#1f2937]"
-                                                            }`}
-                                                        >
-                                                            <div className={`mb-1 text-xs font-bold uppercase tracking-[0.18em] ${isMine ? "text-white/70" : "text-[#6b7280]"}`}>
-                                                                {isMine ? t("You") : displayName}
-                                                            </div>
-                                                            <div className="font-semibold text-sm sm:text-base mb-1">
-                                                                🎮 {t("Game invite")}
-                                                            </div>
-                                                            <div className={`flex gap-2 text-[11px] mb-2 ${isMine ? "text-white/70" : "text-[#6b7280]"}`}>
-                                                                <span>{invite.winningScore !== null ? `${invite.winningScore} ${t("goals")}` : "∞ " + t("goals")}</span>
-                                                                <span>·</span>
-                                                                <span>{invite.duration !== null ? `${invite.duration}s` : "∞"}</span>
-                                                                <span>·</span>
-                                                                <span>{t(THEMES.find(th => th.id === invite.themeId)?.nameKey ?? THEME_DEFAULT.nameKey, invite.themeId)}</span>
-                                                            </div>
-                                                            {resolvedInviteIds.has(invite.matchId) ? (
-                                                                <div className="text-xs text-gray-400 italic mt-1">{t("Invite expired", "Invitation expirée")}</div>
-                                                            ) : isMine ? (
-                                                                <div className="text-xs text-white/80">{t("Waiting for opponent...")}</div>
-                                                            ) : (
-                                                                <div className="flex flex-wrap gap-2 mt-2">
-                                                                    <button
-                                                                        type="button"
-                                                                        disabled={busyInviteId === invite.matchId}
-                                                                        onClick={() => handleAcceptInvite(invite)}
-                                                                        className="rounded-xl border-2 border-[#1f2937] bg-[#4AD95A] px-3 py-1.5 text-xs font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px] disabled:opacity-60"
-                                                                    >
-                                                                        {t("Accept")}
-                                                                    </button>
-                                                                    <button
-                                                                        type="button"
-                                                                        disabled={busyInviteId === invite.matchId}
-                                                                        onClick={() => handleDeclineInvite(invite)}
-                                                                        className="rounded-xl border-2 border-[#1f2937] bg-white px-3 py-1.5 text-xs font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px] disabled:opacity-60"
-                                                                    >
-                                                                        {t("Reject")}
-                                                                    </button>
-                                                                </div>
-                                                            )}
-                                                            {message.created_at && (
-                                                                <div className={`mt-1 text-[10px] ${isMine ? "text-white/50 text-right" : "text-[#9ca3af]"}`}>
-                                                                    {formatMsgTime(message.created_at)}
-                                                                </div>
-                                                            )}
-                                                        </div>
-                                                    ) : (
-                                                        <div
-                                                            className={`rounded-3xl border-2 border-[#1f2937] px-3 py-2.5 shadow-[4px_4px_0_#1f2937] sm:px-4 sm:py-3 ${
-                                                                isMine ? "bg-[#1f2937] text-white" : "bg-white text-[#1f2937]"
-                                                            }`}
-                                                        >
-                                                            <div className={`mb-1 text-xs font-bold uppercase tracking-[0.18em] ${isMine ? "text-white/70" : "text-[#6b7280]"}`}>
-                                                                {isMine ? t("You") : displayName}
-                                                            </div>
-                                                            <div className="break-words text-sm leading-6 sm:text-[0.95rem]">{message.content}</div>
-                                                            {message.created_at && (
-                                                                <div className={`mt-1 text-[10px] ${isMine ? "text-white/50 text-right" : "text-[#9ca3af]"}`}>
-                                                                    {formatMsgTime(message.created_at)}
-                                                                </div>
-                                                            )}
-                                                        </div>
-                                                    )}
-                                                    {seenByPartner && (
-                                                        <div className="text-[10px] text-blue-600 self-end font-semibold uppercase tracking-wider">
-                                                            ✓✓ {t("Seen")}
-                                                        </div>
-                                                    )}
-                                                </div>
-
-                                                {isMine && (
-                                                    <div className="shrink-0 w-9 sm:w-10">
-                                                        {showAvatar ? (
-                                                            <button
-                                                                type="button"
-                                                                onClick={() => openProfile(message.sender_user_id)}
-                                                                aria-label="Open your profile"
-                                                            >
-                                                                <img
-                                                                    src={avatarUrl}
-                                                                    alt={displayName}
-                                                                    onError={handleAvatarError}
-                                                                    className="h-9 w-9 rounded-full border-2 border-[#1f2937] object-cover sm:h-10 sm:w-10"
-                                                                />
-                                                            </button>
-                                                        ) : null}
-                                                    </div>
-                                                )}
-                                            </article>
-                                        );
-                                    })}
-                                </div>
-                            ))}
-
-                            {/* Typing indicator */}
-                            {typingUser && isMember && (
-                                <div className="flex items-center gap-2 text-xs text-[#6b7280] pl-2">
-                                    <span className="inline-flex gap-1">
-                                        <span className="inline-block w-1.5 h-1.5 rounded-full bg-[#6b7280] animate-bounce" style={{ animationDelay: "0ms" }} />
-                                        <span className="inline-block w-1.5 h-1.5 rounded-full bg-[#6b7280] animate-bounce" style={{ animationDelay: "150ms" }} />
-                                        <span className="inline-block w-1.5 h-1.5 rounded-full bg-[#6b7280] animate-bounce" style={{ animationDelay: "300ms" }} />
-                                    </span>
-                                    <span>{(profiles[typingUser.userId]?.display_name || typingUser.username || "Someone")} {t("is typing...")}</span>
-                                </div>
-                            )}
-
-                            {systemEvents.map((event) => (
-                                <div
-                                    key={event.id}
-                                    className="mx-auto max-w-[92%] rounded-2xl border-2 border-dashed border-[#1f2937]/25 bg-white/70 px-4 py-2 text-center text-xs font-medium uppercase tracking-[0.16em] text-[#6b7280]"
-                                >
-                                    {event.text}
-                                </div>
-                            ))}
-
-                            <div ref={messageEndRef} />
-                        </div>
-                    </div>
-
-                    {/* Input area — hidden when no room selected */}
-                    <div className={`border-t-4 border-[#1f2937] bg-white/80 px-3 py-3 backdrop-blur-sm min-[481px]:px-4 ${!selectedRoom ? "hidden" : ""}`}>
-                        <form
-                            className="flex items-center gap-2"
-                            onSubmit={(event) => {
-                                event.preventDefault();
-                                if (!isMember) return;
-                                handleSendMessage();
-                            }}
-                        >
-                            {/* Current user avatar */}
-                            {currentUserId > 0 && (
-                                <img
-                                    src={renderAvatar(currentUserId)}
-                                    alt="You"
-                                    onError={handleAvatarError}
-                                    className="shrink-0 h-8 w-8 rounded-full border-2 border-[#1f2937] object-cover"
-                                />
-                            )}
-                            <input
-                                type="text"
-                                placeholder={isMember ? t("Write a message...") : t("Join the channel to chat")}
-                                value={messageText}
-                                onChange={(e) => {
-                                    setMessageText(e.target.value);
-                                    if (isMember && e.target.value.length > 0) handleTypingPing();
-                                }}
-                                disabled={!isMember}
-                                className="min-w-0 flex-1 rounded-2xl border-2 border-[#1f2937] bg-white px-3 py-2.5 text-sm outline-none transition focus:border-blue-500 disabled:cursor-not-allowed disabled:opacity-60"
-                            />
-                            {/* Game invite button inline — DMs only */}
-                            {isSelectedRoomDm && (
-                                <button
-                                    type="button"
-                                    onClick={handleSendInvite}
-                                    disabled={isInviting}
-                                    title={isInviting ? t("Sending...") : t("Invite to game")}
-                                    className="shrink-0 rounded-2xl border-2 border-[#1f2937] bg-[#facc15] px-2.5 py-2.5 text-base font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px] disabled:cursor-not-allowed disabled:opacity-60"
-                                >
-                                    🎮
-                                </button>
-                            )}
-                            <button
-                                type="submit"
-                                disabled={selectedRoomId == null || isSubmittingMessage || !isMember}
-                                className="shrink-0 rounded-2xl border-2 border-[#1f2937] bg-[#1f2937] px-3 py-2.5 text-sm font-semibold text-white shadow-[3px_3px_0_#1f2937] transition hover:translate-x-[1px] hover:translate-y-[1px] disabled:cursor-not-allowed disabled:opacity-60"
-                            >
-                                {isSubmittingMessage ? "..." : "→"}
-                            </button>
-                        </form>
-                    </div>
+                    {/* Input area */}
+                    <MessageInput
+                        selectedRoom={Boolean(selectedRoom)}
+                        isMember={isMember}
+                        selectedRoomId={selectedRoomId}
+                        isSelectedRoomDm={isSelectedRoomDm}
+                        messageText={messageText}
+                        setMessageText={setMessageText}
+                        isSubmittingMessage={isSubmittingMessage}
+                        currentUserId={currentUserId}
+                        isInviting={invite.isInviting}
+                        renderAvatar={renderAvatar}
+                        handleAvatarError={handleAvatarError}
+                        handleSendMessage={handleSendMessage}
+                        handleSendInvite={invite.handleSendInvite}
+                        handleTypingPing={handleTypingPing}
+                    />
                 </main>
             </div>
 
             {/* ── Modal sélection inviteur ── */}
-            {showInviteModal && (
-                <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/60" onClick={() => setShowInviteModal(false)}>
-                    <div className="w-[min(92vw,22rem)] rounded-3xl border-4 border-[#1f2937] bg-[#f5efe2] p-6 shadow-[8px_8px_0_#1f2937]" onClick={e => e.stopPropagation()}>
-                        <div className="mb-4 text-center text-base font-bold uppercase tracking-widest text-[#1f2937]">🎮 {t("Game invite")}</div>
-
-                        {/* Nation picker */}
-                        <div className="mb-4 flex flex-col items-center gap-2">
-                            <span className="text-xs font-bold uppercase tracking-widest text-[#6b7280]">{t("Your character")}</span>
-                            <div className="flex items-center gap-3">
-                                <button type="button" onClick={() => { const i = CHARACTERS.indexOf(inviteNation as typeof CHARACTERS[number]); setInviteNation(CHARACTERS[(i - 1 + CHARACTERS.length) % CHARACTERS.length]); }} className="rounded-xl border-2 border-[#1f2937] bg-white px-3 py-1.5 text-lg font-bold shadow-[2px_2px_0_#1f2937]">◀</button>
-                                <div className="flex flex-col items-center gap-1 w-24">
-                                    <img src={`/assets/perso/faces/${inviteNation.toLowerCase()}-face.png`} alt={inviteNation} className="w-14 h-14 object-contain" />
-                                    <span className="text-xs font-semibold text-[#1f2937]">{inviteNation}</span>
-                                </div>
-                                <button type="button" onClick={() => { const i = CHARACTERS.indexOf(inviteNation as typeof CHARACTERS[number]); setInviteNation(CHARACTERS[(i + 1) % CHARACTERS.length]); }} className="rounded-xl border-2 border-[#1f2937] bg-white px-3 py-1.5 text-lg font-bold shadow-[2px_2px_0_#1f2937]">▶</button>
-                            </div>
-                        </div>
-
-                        {/* Score */}
-                        <div className="mb-3 flex items-center justify-between gap-2">
-                            <span className="text-xs font-bold uppercase tracking-widest text-[#6b7280]">{t("Goals")}</span>
-                            <div className="flex gap-1.5">
-                                {SCORE_OPTIONS.map(opt => (
-                                    <button key={String(opt)} type="button" onClick={() => setInviteScore(opt)} className={`rounded-xl border-2 border-[#1f2937] px-3 py-1 text-xs font-bold shadow-[2px_2px_0_#1f2937] transition ${inviteScore === opt ? "bg-[#facc15] text-[#1f2937]" : "bg-white text-[#1f2937]"}`}>{opt === null ? "∞" : opt}</button>
-                                ))}
-                            </div>
-                        </div>
-
-                        {/* Timer */}
-                        <div className="mb-3 flex items-center justify-between gap-2">
-                            <span className="text-xs font-bold uppercase tracking-widest text-[#6b7280]">{t("Time")}</span>
-                            <div className="flex gap-1.5">
-                                {TIMER_OPTIONS.map(opt => (
-                                    <button key={String(opt)} type="button" onClick={() => setInviteDuration(opt)} className={`rounded-xl border-2 border-[#1f2937] px-3 py-1 text-xs font-bold shadow-[2px_2px_0_#1f2937] transition ${inviteDuration === opt ? "bg-[#facc15] text-[#1f2937]" : "bg-white text-[#1f2937]"}`}>{opt === null ? "∞" : `${opt}s`}</button>
-                                ))}
-                            </div>
-                        </div>
-
-                        {/* Theme */}
-                        <div className="mb-5 flex items-center justify-between gap-2">
-                            <span className="text-xs font-bold uppercase tracking-widest text-[#6b7280]">{t("Theme")}</span>
-                            <div className="flex gap-1.5">
-                                {THEMES.map(th => (
-                                    <button key={th.id} type="button" onClick={() => setInviteThemeId(th.id)} className={`rounded-xl border-2 border-[#1f2937] px-3 py-1 text-xs font-bold shadow-[2px_2px_0_#1f2937] transition ${inviteThemeId === th.id ? "bg-[#facc15] text-[#1f2937]" : "bg-white text-[#1f2937]"}`}>{t(th.nameKey, th.id)}</button>
-                                ))}
-                            </div>
-                        </div>
-
-                        <div className="flex gap-2">
-                            <button type="button" onClick={() => setShowInviteModal(false)} className="flex-1 rounded-2xl border-2 border-[#1f2937] bg-white py-2 text-sm font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937]">{t("Cancel")}</button>
-                            <button type="button" onClick={handleConfirmInvite} disabled={isInviting} className="flex-1 rounded-2xl border-2 border-[#1f2937] bg-[#4AD95A] py-2 text-sm font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] disabled:opacity-60">{t("Send invite")}</button>
-                        </div>
-                    </div>
-                </div>
-            )}
+            <SendInviteModal
+                showInviteModal={invite.showInviteModal}
+                setShowInviteModal={invite.setShowInviteModal}
+                inviteNation={invite.inviteNation}
+                setInviteNation={invite.setInviteNation}
+                inviteScore={invite.inviteScore}
+                setInviteScore={invite.setInviteScore}
+                inviteDuration={invite.inviteDuration}
+                setInviteDuration={invite.setInviteDuration}
+                inviteThemeId={invite.inviteThemeId}
+                setInviteThemeId={invite.setInviteThemeId}
+                isInviting={invite.isInviting}
+                handleConfirmInvite={invite.handleConfirmInvite}
+            />
 
             {/* ── Modal sélection accepteur ── */}
-            {pendingAccept && (
-                <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/60" onClick={() => setPendingAccept(null)}>
-                    <div className="w-[min(92vw,22rem)] rounded-3xl border-4 border-[#1f2937] bg-[#f5efe2] p-6 shadow-[8px_8px_0_#1f2937]" onClick={e => e.stopPropagation()}>
-                        <div className="mb-4 text-center text-base font-bold uppercase tracking-widest text-[#1f2937]">🎮 {t("Game invite")}</div>
-
-                        {/* Nation picker */}
-                        <div className="mb-3 flex flex-col items-center gap-2">
-                            <span className="text-xs font-bold uppercase tracking-widest text-[#6b7280]">{t("Your character")}</span>
-                            <div className="flex items-center gap-3">
-                                <button type="button" onClick={() => { const i = CHARACTERS.indexOf(acceptNation as typeof CHARACTERS[number]); setAcceptNation(CHARACTERS[(i - 1 + CHARACTERS.length) % CHARACTERS.length]); }} className="rounded-xl border-2 border-[#1f2937] bg-white px-3 py-1.5 text-lg font-bold shadow-[2px_2px_0_#1f2937]">◀</button>
-                                <div className="flex flex-col items-center gap-1 w-24">
-                                    <img src={`/assets/perso/faces/${acceptNation.toLowerCase()}-face.png`} alt={acceptNation} className="w-14 h-14 object-contain" />
-                                    <span className="text-xs font-semibold text-[#1f2937]">{acceptNation}</span>
-                                </div>
-                                <button type="button" onClick={() => { const i = CHARACTERS.indexOf(acceptNation as typeof CHARACTERS[number]); setAcceptNation(CHARACTERS[(i + 1) % CHARACTERS.length]); }} className="rounded-xl border-2 border-[#1f2937] bg-white px-3 py-1.5 text-lg font-bold shadow-[2px_2px_0_#1f2937]">▶</button>
-                            </div>
-                        </div>
-
-                        {/* Paramètres imposés par l'inviteur (lecture seule) */}
-                        <div className="mb-5 flex items-center justify-between gap-2 text-xs text-[#6b7280]">
-                            <span>{pendingAccept.winningScore !== null ? `${pendingAccept.winningScore} ${t("goals")}` : "∞ " + t("goals")}</span>
-                            <span>{pendingAccept.duration !== null ? `${pendingAccept.duration}s` : "∞"}</span>
-                            <span>{t(THEMES.find(th => th.id === pendingAccept.themeId)?.nameKey ?? THEME_DEFAULT.nameKey, pendingAccept.themeId)}</span>
-                        </div>
-
-                        <div className="flex gap-2">
-                            <button type="button" onClick={() => { handleDeclineInvite(pendingAccept); setPendingAccept(null); }} className="flex-1 rounded-2xl border-2 border-[#1f2937] bg-white py-2 text-sm font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937]">{t("Decline")}</button>
-                            <button type="button" onClick={handleConfirmAccept} disabled={busyInviteId === pendingAccept.matchId} className="flex-1 rounded-2xl border-2 border-[#1f2937] bg-[#4AD95A] py-2 text-sm font-semibold text-[#1f2937] shadow-[2px_2px_0_#1f2937] disabled:opacity-60">{t("Accept")}</button>
-                        </div>
-                    </div>
-                </div>
-            )}
+            <AcceptInviteModal
+                pendingAccept={invite.pendingAccept}
+                setPendingAccept={invite.setPendingAccept}
+                acceptNation={invite.acceptNation}
+                setAcceptNation={invite.setAcceptNation}
+                busyInviteId={invite.busyInviteId}
+                handleConfirmAccept={invite.handleConfirmAccept}
+                handleDeclineInvite={invite.handleDeclineInvite}
+            />
         </div>
     );
 }
-
